refactor(65-08): icons as committed PNGs, drop procedural rasterizer
- add raw/{chevron_down,arrow_right,caret}.png (white straight-alpha, builder premultiplies)
- add tools/gen_icons.c manual dev tool (regenerates the PNGs; not in any CMake target)
- build_packs.c: remove icon_tri_cover/icon_raster_tri/add_icon_tri + add_raw icon calls;
  pack the 3 PNGs as normal sprites with identical names (region constants unchanged); _white stays
